// src/app/omsorgspenger-kronisk-sykt-barn/types/EksisterendeOMPKSSoknaderActionKeys.ts
 import/prefer-default-export
export enum EksisterendeOMPKSSoknaderActionKeys {
    EKSISTERENDE_OMP_KS_SOKNADER_SET = 'PUNCH_EKSISTERENDE_OMP_KS_SOKNADER_SET',
    EKSISTERENDE_OMP_KS_SOKNADER_LOAD = 'PUNCH_EKSISTERENDE_OMP_KS_SOKNADER_LOAD',
    EKSISTERENDE_OMP_KS_SOKNADER_REQUEST_ERROR = 'PUNCH_EKSISTERENDE_OMP_KS_SOKNADER_REQUEST_ERROR',

    EKSISTERENDE_OMP_KS_SOKNAD_OPEN = 'PUNCH_EKSISTERENDE_OMP_KS_SOKNAD_OPEN',
    EKSISTERENDE_OMP_KS_SOKNAD_CLOSE = 'PUNCH_EKSISTERENDE_OMP_KS_SOKNAD_CLOSE',
    EKSISTERENDE_OMP_KS_SOKNAD_CHOOSE = 'PUNCH_EKSISTERENDE_OMP_KS_SOKNAD_CHOOSE',
    EKSISTERENDE_OMP_KS_SOKNAD_UNDO_CHOICE = 'PUNCH_EKSISTERENDE_OMP_KS_SOKNAD_UNDO_CHOICE',

    OMP_KS_SOKNAD_CREATE_REQUEST = 'PUNCH_EKSISTERENDE_OMP_KS_SOKNAD_CREATE_REQUEST',
    OMP_KS_SOKNAD_CREATE_ERROR = 'PUNCH_EKSISTERENDE_OMP_KS_SOKNAD_CREATE_ERROR',
    OMP_KS_SOKNAD_CREATE_SUCCESS = 'PUNCH_EKSISTERENDE_OMP_KS_SOKNAD_CREATE_SUCCESS',

    OMP_KS_SOKNADID_RESET = 'PUNCH_EKSTISTERENDE_OMP_KS_SOKNADID_RESET',
}